% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courses.R
\docType{data}
\name{courses}
\alias{courses}
\title{Harvard Course Information}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26477 rows and 6 columns.}
\usage{
courses
}
\description{
Thanks to Registrar Mike Burke for providing this information.
}
\keyword{datasets}
