% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elections.R
\docType{data}
\name{elections}
\alias{elections}
\title{County Presidential Election Results from 1960 through 2012}
\format{A tibble with 43,494 rows and 7 variables:
\describe{
  \item{year}
  \item{state}
  \item{county}
  \item{rep}{number of votes cast for the Republican candidate}
  \item{dem}{number of votes cast for the Democratic candidate}
  \item{other}{number of votes cast for other candidates}
  \item{rep_per}{percentage of all votes cast for the Republican candidate}
}}
\source{
\url{http://qss.princeton.press/}
}
\usage{
elections
}
\description{
A dataset presidential election results for each US county.
}
\keyword{datasets}
