% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qscores.R
\docType{data}
\name{qscores}
\alias{qscores}
\title{Scores from the Harvard Q Guide}
\format{A tibble with 141 observations and 6 variables:
\describe{
  \item{names}{Course full name}
  \item{course}{Course number}
  \item{department}{Course department}
  \item{overall}{average overall student evaluation on a scale of 1 to 5}
  \item{workload}{average weekly workload in hours}
  \item{number}{total number of students filling out evaluation forms}
}}
\source{
\url{http://q-guide-plus.appspot.com/}
}
\usage{
qscores
}
\description{
Thanks to Benny Chang '22 for this data.
}
\keyword{datasets}
