% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes_2016.R
\docType{data}
\name{votes_2016}
\alias{votes_2016}
\title{County Presidential Election Results for 2016}
\format{A data frame with 3,112 rows and 39 variables.}
\source{
Courtesy of Xiang Zhou and Gov 2000.
}
\usage{
votes_2016
}
\description{
A dataset presidential election votes for each US county.
}
\keyword{datasets}
