% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qscores.R
\docType{data}
\name{qscores}
\alias{qscores}
\title{Scores from the Harvard Q-Guide}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4409 rows and 10 columns.}
\usage{
qscores
}
\description{
Thanks to Benny Chang '22 for this data
}
\keyword{datasets}
